+++ purchases/src/test/java/com/revenuecat/purchases/PurchasesTest.kt
@@ -2764,13 +2764,15 @@ class PurchasesTest {
     }
 
     @Test
-    fun `syncing a transaction posts normalized purchase data to backend`() {
+    fun `syncing an Amazon transaction posts normalized purchase data to backend`() {
         purchases.finishTransactions = false
 
         val skuParent = "sub"
         val skuTerm = "sub.monthly"
         val purchaseToken = "crazy_purchase_token"
         val amazonUserID = "amazon_user_id"
+        val price = 10.40
+        val currencyCode = "USD"
 
         every {
             mockBillingAbstract.normalizePurchaseData(
@@ -2793,11 +2795,15 @@ class PurchasesTest {
         purchases.syncObserverModeAmazonPurchase(
             productID = skuParent,
             receiptId = purchaseToken,
-            amazonUserID = amazonUserID
+            amazonUserID = amazonUserID,
+            price = price,
+            isoCurrencyCode = currencyCode
         )
 
         val productInfo = ReceiptInfo(
-            productID = skuTerm
+            productID = skuTerm,
+            price = price,
+            currency = currencyCode
         )
         verify(exactly = 1) {
             mockBackend.postReceiptData(
@@ -2815,13 +2821,15 @@ class PurchasesTest {
     }
 
     @Test
-    fun `syncing a transaction respects allow sharing account settings`() {
+    fun `syncing an Amazon transaction respects allow sharing account settings`() {
         purchases.finishTransactions = false
 
         val skuParent = "sub"
         val skuTerm = "sub.monthly"
         val purchaseToken = "crazy_purchase_token"
         val amazonUserID = "amazon_user_id"
+        val price = 10.40
+        val currencyCode = "USD"
         purchases.allowSharingPlayStoreAccount = true
 
         every {
@@ -2845,11 +2853,15 @@ class PurchasesTest {
         purchases.syncObserverModeAmazonPurchase(
             productID = skuParent,
             receiptId = purchaseToken,
-            amazonUserID = amazonUserID
+            amazonUserID = amazonUserID,
+            price = price,
+            isoCurrencyCode = currencyCode
         )
 
         val productInfo = ReceiptInfo(
-            productID = skuTerm
+            productID = skuTerm,
+            price = price,
+            currency = currencyCode
         )
         verify(exactly = 1) {
             mockBackend.postReceiptData(
@@ -2867,13 +2879,15 @@ class PurchasesTest {
     }
 
     @Test
-    fun `syncing a transaction never consumes it`() {
+    fun `syncing an Amazon transaction never consumes it`() {
         purchases.finishTransactions = false
 
         val skuParent = "sub"
         val skuTerm = "sub.monthly"
         val purchaseToken = "crazy_purchase_token"
         val amazonUserID = "amazon_user_id"
+        val price = 10.40
+        val currencyCode = "USD"
         purchases.allowSharingPlayStoreAccount = true
 
         var capturedLambda: ((String) -> Unit)? = null
@@ -2898,11 +2912,15 @@ class PurchasesTest {
         purchases.syncObserverModeAmazonPurchase(
             productID = skuParent,
             receiptId = purchaseToken,
-            amazonUserID = amazonUserID
+            amazonUserID = amazonUserID,
+            price = price,
+            isoCurrencyCode = currencyCode
         )
 
         val productInfo = ReceiptInfo(
-            productID = skuTerm
+            productID = skuTerm,
+            price = price,
+            currency = currencyCode
         )
         verify(exactly = 1) {
             mockBackend.postReceiptData(
@@ -2925,13 +2943,15 @@ class PurchasesTest {
     }
 
     @Test
-    fun `transaction is not synced again if it was already synced`() {
+    fun `Amazon transaction is not synced again if it was already synced`() {
         purchases.finishTransactions = false
 
         val skuParent = "sub"
         val skuTerm = "sub.monthly"
         val purchaseToken = "crazy_purchase_token"
         val amazonUserID = "amazon_user_id"
+        val price = 10.40
+        val currencyCode = "USD"
 
         every {
             mockCache.getPreviouslySentHashedTokens()
@@ -2954,11 +2974,15 @@ class PurchasesTest {
         purchases.syncObserverModeAmazonPurchase(
             productID = skuParent,
             receiptId = purchaseToken,
-            amazonUserID = amazonUserID
+            amazonUserID = amazonUserID,
+            price = price,
+            isoCurrencyCode = currencyCode
         )
 
         val productInfo = ReceiptInfo(
-            productID = skuTerm
+            productID = skuTerm,
+            price = price,
+            currency = currencyCode
         )
         verify(exactly = 1) {
             mockBackend.postReceiptData(
@@ -2985,7 +3009,9 @@ class PurchasesTest {
         purchases.syncObserverModeAmazonPurchase(
             productID = skuParent,
             receiptId = purchaseToken,
-            amazonUserID = amazonUserID
+            amazonUserID = amazonUserID,
+            price = price,
+            isoCurrencyCode = currencyCode
         )
 
         verify(exactly = 1) {
@@ -3012,13 +3038,15 @@ class PurchasesTest {
     }
 
     @Test
-    fun `syncing a transaction sends subscriber attributes`() {
+    fun `syncing an Amazon transaction sends subscriber attributes`() {
         purchases.finishTransactions = false
 
         val skuParent = "sub"
         val skuTerm = "sub.monthly"
         val purchaseToken = "crazy_purchase_token"
         val amazonUserID = "amazon_user_id"
+        val price = 10.40
+        val currencyCode = "USD"
         val subscriberAttributeKey = "favorite_cat"
         val subscriberAttributeValue = "gardfield"
         val subscriberAttribute = SubscriberAttribute(subscriberAttributeKey, subscriberAttributeValue)
@@ -3051,11 +3079,15 @@ class PurchasesTest {
         purchases.syncObserverModeAmazonPurchase(
             productID = skuParent,
             receiptId = purchaseToken,
-            amazonUserID = amazonUserID
+            amazonUserID = amazonUserID,
+            price = price,
+            isoCurrencyCode = currencyCode
         )
 
         val productInfo = ReceiptInfo(
-            productID = skuTerm
+            productID = skuTerm,
+            price = price,
+            currency = currencyCode
         )
         verify(exactly = 1) {
             mockBackend.postReceiptData(
@@ -3075,6 +3107,112 @@ class PurchasesTest {
             mockBillingAbstract.consumeAndSave(any(), any())
         }
         assertThat(capturedLambda).isNotNull
+    }
+
+    @Test
+    fun `syncing an Amazon transaction without price nor currency code posts purchase data to backend`() {
+        purchases.finishTransactions = false
+
+        val skuParent = "sub"
+        val skuTerm = "sub.monthly"
+        val purchaseToken = "crazy_purchase_token"
+        val amazonUserID = "amazon_user_id"
+
+        every {
+            mockBillingAbstract.normalizePurchaseData(
+                productID = skuParent,
+                purchaseToken = purchaseToken,
+                storeUserID = amazonUserID,
+                captureLambda(),
+                any()
+            )
+        } answers {
+            lambda<(String) -> Unit>().captured.also {
+                it.invoke(skuTerm)
+            }
+        }
+
+        every {
+            mockCache.getPreviouslySentHashedTokens()
+        } returns setOf()
+
+        purchases.syncObserverModeAmazonPurchase(
+            productID = skuParent,
+            receiptId = purchaseToken,
+            amazonUserID = amazonUserID,
+            price = null,
+            isoCurrencyCode = null
+        )
+
+        val productInfo = ReceiptInfo(productID = skuTerm)
+        verify(exactly = 1) {
+            mockBackend.postReceiptData(
+                purchaseToken = purchaseToken,
+                appUserID = appUserId,
+                isRestore = false,
+                observerMode = true,
+                subscriberAttributes = emptyMap(),
+                receiptInfo = productInfo,
+                storeAppUserID = amazonUserID,
+                onSuccess = any(),
+                onError = any()
+            )
+        }
+    }
+
+    @Test
+    fun `syncing an Amazon transaction with zero price posts correct purchase data to backend`() {
+        purchases.finishTransactions = false
+
+        val skuParent = "sub"
+        val skuTerm = "sub.monthly"
+        val purchaseToken = "crazy_purchase_token"
+        val amazonUserID = "amazon_user_id"
+
+        every {
+            mockBillingAbstract.normalizePurchaseData(
+                productID = skuParent,
+                purchaseToken = purchaseToken,
+                storeUserID = amazonUserID,
+                captureLambda(),
+                any()
+            )
+        } answers {
+            lambda<(String) -> Unit>().captured.also {
+                it.invoke(skuTerm)
+            }
+        }
+
+        every {
+            mockCache.getPreviouslySentHashedTokens()
+        } returns setOf()
+
+        purchases.syncObserverModeAmazonPurchase(
+            productID = skuParent,
+            receiptId = purchaseToken,
+            amazonUserID = amazonUserID,
+            price = 0.0,
+            isoCurrencyCode = null
+        )
+
+        val productInfo = ReceiptInfo(
+            productID = skuTerm,
+            currency = null,
+            price = null
+        )
+        verify(exactly = 1) {
+            mockBackend.postReceiptData(
+                purchaseToken = purchaseToken,
+                appUserID = appUserId,
+                isRestore = false,
+                observerMode = true,
+                subscriberAttributes = emptyMap(),
+                receiptInfo = productInfo,
+                storeAppUserID = amazonUserID,
+                onSuccess = any(),
+                onError = any()
+            )
+        }
     }
 
     @Test