+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 5.0.0-amazon.alpha.6.billing3.iap2
+
+- Fix product price and currency not being reported in Amazon Appstore observer mode. `syncObserverModeAmazonPurchase`
+  now accepts an `isoCurrencyCode` and a `price` to make sure we correctly track prices.
+  https://github.com/RevenueCat/purchases-android/pull/519
+
 ## 5.0.0-amazon.alpha.5.billing3.iap2
 
 - Skip registering listener in Amazon observer mode